correction chargement img à la création du post

+++ backend/controllers/post.js
@@ -6,7 +6,25 @@ const fs = require('fs'); // pour les fichiers img
 
 // .save() n'est pas nécessaire en sequelize
 
+/*
+exports.createPost = (req, res) => {
+  return Post.create({
+    title: res.title, 
+    description: res.description,
+    userId: req.userId
+  })
+  .then((post) => {
+    return post; 
+  })
+  .catch((err) => {
+    console.log(err);
+      
+    });
+  };
+*/
+
 exports.createPost = (req, res, next) => {
+  //ici faire les console.log() -> regarder dans le terminal du back
   if(!req.body.title) {
     res.status(400).json({
       message: "Le contenu ne peut pas être vide"
@@ -17,16 +35,15 @@ exports.createPost = (req, res, next) => {
     {
       ...req.body, 
       imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-    } : { ...req.body };
+    } : { 
+      ...req.body,
+    };
   Post.create(postObject)
   .then((post) => {
-    //post.setUser(req.user); // userId n'est pas récupéré 
-    //cf. doc juste fait userId = userId et envoyer manuellement l'id utilisateur du front...
+    post.setUser(req.user);     
     return res.status(201).json(post); 
-    // vérifier qu'on pourra bien récupérer le photo de profil de l'utilisateur
   })
   .catch((err) => {
-    console.log(req.body.imageUrl)
     return res.status(500).json({
       message: err.message || "Erreur pendant la création du post"
       
@@ -35,7 +52,7 @@ exports.createPost = (req, res, next) => {
 };
 
 exports.findAllPost = (req, res, next) => {
-  /* se faire expliquer ceci ??
+  /* 
   const title = req.query.title;
   var condition = title ? { title: { [Op.like]: `%${title}%`} }: null; */
 
@@ -93,14 +110,26 @@ exports.updatePost = (req, res, next) => {
 exports.deletePost = (req, res, next) => {
   Post.findByPk( req.params.id )
   .then(post => {
-    // on récupére le nom du fichier,
-    //on split la chaine de caractères au niveau de images puis on prend la 2nd valeur
-    const filename = post.imageUrl.split('/images/')[1];
-    // on appelle une fonciton de fs 'unlike' pour supprimer un fichier 
-    //elle prend en paramétre le dossier et le nom de l'image
-    fs.unlink(`images/${filename}`, () => {
-      // dans le callback de fs.unlike, une fois l'image supprimé 
-      //on supprime le reste des info du post
+    if(req.file) {
+      const filename = post.imageUrl.split('/images/')[1];
+      fs.unlink(`images/${filename}`, () => {
+        Post.destroy({where: {id: req.params.id} }) 
+        .then(num => {
+          if(num == 1) {
+            return res.status(200).json({ message: "Post was deleted successfully !" });
+          } else {
+            return res.status(400).json({
+              message: `Can't delete Post with id=${id}. Maybe Post wasn't found`
+            });
+          }
+        })
+        .catch((err) => {
+          return res.status(500).json({
+            message: `Error deleting Post with id=${id}`
+          });
+        });
+      });
+    } else {
       Post.destroy({where: {id: req.params.id} }) 
       .then(num => {
         if(num == 1) {
@@ -116,7 +145,7 @@ exports.deletePost = (req, res, next) => {
           message: `Error deleting Post with id=${id}`
         });
       });
-    });
+    }
   })
   .catch(error => res.status(500).json({  message: "Le post n'existe pas"  }));
 };
